+++ rorapi/common/csv_update.py
@@ -1,355 +1,13 @@
-import jsonschema
-import requests
 import copy
-import csv
-import json
-import io
-import os
-import re
-import shutil
-import urllib
-from datetime import datetime
-from iso639 import Lang
-from rest_framework.exceptions import ParseError
-from rest_framework.parsers import JSONParser
-from rest_framework.renderers import JSONRenderer
-from rorapi.common.models import Errors
-import update_address as ua
-from rorapi.settings import DATA
-from rorapi.v2.models import (
-    Organization as OrganizationV2,
-    ListResult as ListResultV2
-)
+from rorapi.common.record_utils import *
+from rorapi.v2.record_constants import *
+from rorapi.common.csv_utils import *
 from rorapi.v2.serializers import (
     OrganizationSerializer as OrganizationSerializerV2
 )
-from rorapi.common.queries import get_ror_id, retrieve_organization
+from rorapi.common.queries import retrieve_organization
 from rorapi.common.serializers import ErrorsSerializer
-
-from rorapi.management.commands.generaterorid import check_ror_id
-
-ADMIN = {
-    "created": {
-        "date": "",
-        "schema_version": "2.0"
-    },
-    "last_modified": {
-        "date": "",
-        "schema_version": "2.0"
-    }
-}
-
-LAST_MOD = {
-    "date": "",
-    "schema_version": "2.0"
-}
-
-OPTIONAL_FIELD_DEFAULTS = {
-    "domains": [],
-    "established": None,
-    "external_ids": [],
-    "links": [],
-    "relationships": []
-}
-
-UPDATE_ACTIONS = {
-    "ADD": "add",
-    "DELETE": "delete",
-    "REPLACE": "replace"
-}
-
-UPDATE_ACTIONS_MULTI = [UPDATE_ACTIONS["ADD"], UPDATE_ACTIONS["DELETE"], UPDATE_ACTIONS["REPLACE"]]
-
-UPDATE_ACTIONS_SINGLE = [UPDATE_ACTIONS["DELETE"], UPDATE_ACTIONS["REPLACE"]]
-
-NO_DELETE_FIELDS = ["id", "locations.geonames_id", "names.types.ror_display", "status", "types"]
-
-CSV_REQUIRED_FIELDS_ACTIONS = {
-    "id": None,
-    "domains": UPDATE_ACTIONS_MULTI,
-    "established": UPDATE_ACTIONS_SINGLE,
-    "external_ids.type.fundref.all": UPDATE_ACTIONS_MULTI,
-    "external_ids.type.fundref.preferred": UPDATE_ACTIONS_SINGLE,
-    "external_ids.type.grid.all": UPDATE_ACTIONS_MULTI,
-    "external_ids.type.grid.preferred": UPDATE_ACTIONS_SINGLE,
-    "external_ids.type.isni.all": UPDATE_ACTIONS_MULTI,
-    "external_ids.type.isni.preferred": UPDATE_ACTIONS_SINGLE,
-    "external_ids.type.wikidata.all": UPDATE_ACTIONS_MULTI,
-    "external_ids.type.wikidata.preferred": UPDATE_ACTIONS_SINGLE,
-    "links.type.website": UPDATE_ACTIONS_MULTI,
-    "links.type.wikipedia": UPDATE_ACTIONS_MULTI,
-    "locations.geonames_id": UPDATE_ACTIONS_MULTI,
-    "names.types.acronym": UPDATE_ACTIONS_MULTI,
-    "names.types.alias": UPDATE_ACTIONS_MULTI,
-    "names.types.label": UPDATE_ACTIONS_MULTI,
-    "names.types.ror_display": [UPDATE_ACTIONS["REPLACE"]],
-    "status": [UPDATE_ACTIONS["REPLACE"]],
-    "types": UPDATE_ACTIONS_MULTI
-}
-
-V2_TEMPLATE = {
-    "locations": [],
-    "established": None,
-    "external_ids": [],
-    "id": "",
-    "domains": [],
-    "links": [],
-    "names": [],
-    "relationships": [],
-    "status": "",
-    "types": [],
-    "admin": {}
-}
-
-V2_EXTERNAL_ID_TYPES = {
-                        "FUNDREF" : "fundref",
-                        "GRID" : "grid",
-                        "ISNI" : "isni",
-                        "WIKIDATA" : "wikidata"
-                    }
-
-V2_LINK_TYPES = {
-                "WEBSITE" : "website",
-                "WIKIPEDIA" : "wikipedia"
-            }
-
-V2_NAME_TYPES = {
-                "ACRONYM" : "acronym",
-                "ALIAS" : "alias",
-                "LABEL" : "label",
-                "ROR_DISPLAY" : "ror_display"
-            }
-
-LANG_DELIMITER = "*"
-
-UPDATE_DELIMITER = "=="
-
-SORT_KEYS = {
-    "domains": None,
-    "external_ids": "type",
-    "links": "type",
-    "locations": "geonames_id",
-    "names": "value",
-    "relationships": "type",
-    "types": None
-}
-
-def update_record(json_input, existing_record):
-    record = copy.deepcopy(existing_record)
-    for k, v in json_input.items():
-        record[k] = copy.deepcopy(v)
-    return record
-
-def update_last_mod(record):
-    record['admin']['last_modified'] = copy.deepcopy(LAST_MOD)
-    record['admin']['last_modified']['date'] = datetime.now().strftime("%Y-%m-%d")
-    return record
-
-def check_optional_fields(record):
-    for k in OPTIONAL_FIELD_DEFAULTS:
-        if k not in record:
-            return True
-    return False
-
-def add_missing_optional_fields(record):
-    for k, v in OPTIONAL_FIELD_DEFAULTS.items():
-        if k not in record:
-            record[k] = v
-    return record
-
-def add_created_last_mod(record):
-    today = datetime.now().strftime("%Y-%m-%d")
-    record['admin'] = copy.deepcopy(ADMIN)
-    record['admin']['created']['date'] = today
-    record['admin']['last_modified']['date'] = today
-    return record
-
-def update_locations(locations):
-    errors = []
-    updated_locations = []
-    for location in locations:
-        if 'geonames_id' in location:
-            try:
-                print(location['geonames_id'])
-                updated_location = ua.new_geonames_v2(str(location['geonames_id']))
-                updated_locations.append(updated_location['location'])
-            except:
-                errors.append("Error retrieving Geonames data for ID {}. Please check that this is a valid Geonames ID".format(location['geonames_id']))
-    return errors, updated_locations
-
-def get_lang_code(lang_string):
-    lang_code = None
-    error = None
-    if len(lang_string) == 2:
-        lang_string = lang_string.lower()
-    else:
-        lang_string = lang_string.title()
-    try:
-        lg = Lang(lang_string)
-        lang_code = lg.pt1
-    except Exception as e:
-        error = e.msg
-    return error, lang_code
-
-def sort_list_fields(v2_record):
-    for field in v2_record:
-        if field in SORT_KEYS:
-            if SORT_KEYS[field] is not None:
-                sort_key = SORT_KEYS[field]
-                sorted_vals = sorted(v2_record[field], key=lambda x: x[sort_key])
-            else:
-                sorted_vals = v2_record[field].sort()
-            v2_record[field] = sorted_vals
-    return v2_record
-
-
-def get_file_from_url(url):
-    rsp = requests.get(url)
-    rsp.raise_for_status()
-    return rsp.json()
-
-def validate_v2(data):
-    errors = []
-    schema = get_file_from_url("https://raw.githubusercontent.com/ror-community/ror-schema/schema-v2/ror_schema_v2_0.json")
-    try:
-        print("validating data:")
-        print(data)
-        jsonschema.validate(data, schema)
-    except jsonschema.ValidationError as error:
-        errors.append(error)
-        print(errors)
-        return Errors(errors), None
-    else:
-        return None, data
-
-def validate_csv(csv_file):
-    errors = []
-    try:
-        read_file = csv_file.read().decode('utf-8')
-        reader = csv.DictReader(io.StringIO(read_file))
-        rowcount = 0
-        for row in reader:
-            rowcount += 1
-        if rowcount > 0:
-            csv_fields = reader.fieldnames
-            missing_fields = []
-            for field in CSV_REQUIRED_FIELDS_ACTIONS.keys():
-                if field not in csv_fields:
-                    missing_fields.append(field)
-            print(missing_fields)
-            if missing_fields:
-                errors.append(f'CSV file is missing columns: {", ".join(missing_fields)}')
-        else:
-            errors.append("CSV file contains no data rows")
-    except IOError as e:
-        errors.append(f"Error parsing CSV file: {e}")
-    print(errors)
-    return errors
-
-
-def new_record_from_json(json_input, version):
-    errors = None
-    valid_data = None
-    new_record = copy.deepcopy(json_input)
-    if check_optional_fields(new_record):
-        new_record = add_missing_optional_fields(new_record)
-    location_errors, updated_locations = update_locations(new_record['locations'])
-    if location_errors:
-        errors = Errors(location_errors)
-    else:
-        new_record['locations'] = updated_locations
-        new_record = add_created_last_mod(new_record)
-        new_ror_id = check_ror_id(version)
-        print("new ror id: " + new_ror_id)
-        new_record['id'] = new_ror_id
-        errors, valid_data = validate_v2(new_record)
-    return errors, valid_data
-
-
-def update_record_from_json(new_json, existing_org):
-    errors = None
-    valid_data = None
-    serializer = OrganizationSerializerV2(existing_org)
-    existing_record = serializer.data
-    updated_record = update_record(new_json, existing_record)
-    location_errors, updated_locations = update_locations(updated_record['locations'])
-    if location_errors:
-        errors = Errors(location_errors)
-    else:
-        updated_record['locations'] = updated_locations
-        errors, valid_data = validate_v2(updated_record)
-    return errors, valid_data
-
-
-def get_action_value(csv_field):
-    action = None
-    value = None
-    if csv_field.lower() == "delete":
-        action = "delete"
-        value = None
-    elif UPDATE_DELIMITER in csv_field:
-        action = csv_field.split(UPDATE_DELIMITER)[0]
-        value = csv_field.split(UPDATE_DELIMITER)[1]
-    else:
-        action = "replace"
-        value = csv_field
-    return action, value
-
-def get_actions_values(csv_field):
-    print("getting actions values:")
-    actions_values = {}
-    if csv_field.lower() == UPDATE_ACTIONS["DELETE"]:
-        actions_values[UPDATE_ACTIONS["DELETE"]] = None
-    elif UPDATE_DELIMITER in csv_field:
-        for ua in list(UPDATE_ACTIONS.values()):
-            print(ua)
-            if ua + UPDATE_DELIMITER in csv_field:
-                print("doing regex:")
-                regex = r"(" + re.escape(
-      ua + UPDATE_DELIMITER) + r")(.*?)(?=$|(add|delete|replace)==)"
-                result = re.search(regex, csv_field)
-                print(result[0])
-                temp_val = result[0].replace(ua + UPDATE_DELIMITER, '')
-                print("temp val:")
-                print(temp_val)
-                actions_values[ua] = [v.strip() for v in temp_val.split(';') if v]
-                #csv_field.replace(result[0], '')
-
-    else:
-        actions_values[UPDATE_ACTIONS["REPLACE"]] = [v.strip() for v in csv_field.split(';') if v]
-    print(actions_values)
-    return actions_values
-
-def validate_csv_row_update_syntax(csv_data):
-    print("validating row")
-    errors = []
-    for k, v in csv_data.items():
-        if UPDATE_DELIMITER in v:
-            print("field:")
-            print(k)
-            print("value:")
-            print(v)
-            actions_values = get_actions_values(v)
-            print("actions values:")
-            print(actions_values)
-            update_actions = list(actions_values.keys())
-            if not update_actions:
-                errors.append("Update delimiter '{}' found in '{}' field but no valid update action found in value {}".format(UPDATE_DELIMITER, k, v))
-            if len(update_actions) > 2:
-                errors.append("{} update actions '{}' found in '{}' field but only 2 are allowed".format(str(len(update_actions)), ", ".join(update_actions), k))
-            if len(update_actions) == 2:
-                if not (UPDATE_ACTIONS['ADD'] and UPDATE_ACTIONS['DELETE']) in update_actions:
-                    errors.append("Invalid combination of update actions '{}' found in '{}' field.".format(", ".join(update_actions), k))
-            disallowed_actions = [ua for ua in update_actions if ua not in CSV_REQUIRED_FIELDS_ACTIONS[k]]
-            print("allowed actions:")
-            print(CSV_REQUIRED_FIELDS_ACTIONS[k])
-            print("disallowed actions:")
-            print(disallowed_actions)
-            if disallowed_actions:
-                errors.append("Invalid update action(s) '{}' found in {} field. Allowed actions for this field are '{}'".format(", ".join(disallowed_actions), k, ", ".join(CSV_REQUIRED_FIELDS_ACTIONS[k])))
-        if v.strip() == UPDATE_ACTIONS['DELETE'].lower() and k in NO_DELETE_FIELDS:
-             errors.append("Invalid update action '{}' in {} field. Cannot remove all values from a required field.".format(UPDATE_ACTIONS['DELETE'], k))
-    return errors
+from rorapi.common.create_update import update_record_from_json
 
 def update_record_from_csv(csv_data, version):
     errors = []
@@ -724,199 +382,3 @@ def update_record_from_csv(csv_data, version):
                 if validation_errors:
                     errors = ErrorsSerializer(validation_errors).data
     return errors, updated_record
-
-
-def new_record_from_csv(csv_data, version):
-    v2_data = copy.deepcopy(V2_TEMPLATE)
-    errors = []
-    #domains
-    if csv_data['domains']:
-        v2_data['domains'] = [d.strip() for d in csv_data['domains'].strip(';').split(';')]
-
-    #established
-    if csv_data['established']:
-        v2_data['established'] = int(csv_data['established'].strip())
-
-    #external ids
-    for k,v in V2_EXTERNAL_ID_TYPES.items():
-        if csv_data['external_ids.type.' + v + '.all']:
-            all_ids = [i.strip() for i in csv_data['external_ids.type.' + v + '.all'].strip(';').split(';')]
-            ext_id_obj = {
-                "type": v,
-                "all": all_ids,
-                "preferred": csv_data['external_ids.type.' + v + '.preferred'].strip() if csv_data['external_ids.type.' + v + '.preferred'] else all_ids[0]
-            }
-            v2_data['external_ids'].append(ext_id_obj)
-
-    #links
-    for k,v in V2_LINK_TYPES.items():
-        if csv_data['links.type.' + v]:
-            for l in csv_data['links.type.' + v].strip(';').split(';'):
-                link_obj = {
-                    "type": v,
-                    "value": l.strip()
-                }
-                v2_data['links'].append(link_obj)
-
-    #locations
-    if csv_data['locations.geonames_id']:
-        geonames_ids = [i.strip() for i in csv_data['locations.geonames_id'].strip(';').split(';')]
-        for geonames_id in geonames_ids:
-            location_obj = {
-                "geonames_id": geonames_id,
-                "geonames_details": {}
-            }
-            v2_data['locations'].append(location_obj)
-
-    #names
-    temp_names = []
-    for k,v in V2_NAME_TYPES.items():
-        if csv_data['names.types.' + v]:
-            for n in csv_data['names.types.' + v].strip(';').split(';'):
-                if LANG_DELIMITER in n:
-                    name_val, lang  = n.split("*")
-                    if lang:
-                        lang_errors, lang_code = get_lang_code(lang.strip())
-                        if lang_errors:
-                            errors.append("Could not convert language value to ISO code: {}".format(lang))
-                else:
-                    name_val = n
-                    lang_code = None
-
-                name_obj = {
-                    "types": [v],
-                    "value": name_val.strip(),
-                    "lang": lang_code
-                }
-                temp_names.append(name_obj)
-    print("temp names 1:")
-    print(temp_names)
-    name_values = [n['value'] for n in temp_names]
-    dup_names = []
-    for n in name_values:
-        if name_values.count(n) > 1:
-            if n not in dup_names:
-                dup_names.append(n)
-    if dup_names:
-        dup_names_objs = []
-        for d in dup_names:
-            types = []
-            for t in temp_names:
-                if t['value'] == d:
-                    types.extend(t['types'])
-            name_obj = {
-                "types": types,
-                "value": d,
-                "lang": None
-            }
-            dup_names_objs.append(name_obj)
-        temp_names = [t for t in temp_names if t['value'] not in dup_names]
-        temp_names.extend(dup_names_objs)
-    print("temp names 2:")
-    print(temp_names)
-    v2_data['names'] = temp_names
-
-    #status
-    if csv_data['status']:
-        v2_data['status'] = csv_data['status'].strip().lower()
-
-    #types
-    if csv_data['types']:
-        v2_data['types'] = [t.strip().lower() for t in csv_data['types'].strip(';').split(';')]
-
-    validation_errors, new_record = new_record_from_json(v2_data, version)
-    if validation_errors:
-        errors = ErrorsSerializer(validation_errors).data
-    return errors, new_record
-
-def save_record_file(ror_id, updated, json_obj, dir_name):
-    dir_path = os.path.join(DATA['DIR'],dir_name)
-    if not os.path.exists(dir_path):
-        os.mkdir(dir_path)
-    subdir = 'updates' if updated else 'new'
-    if not os.path.exists(os.path.join(dir_path, subdir)):
-        os.mkdir(os.path.join(dir_path, subdir))
-    full_path = os.path.join(dir_path, subdir, ror_id.split('https://ror.org/')[1] + '.json')
-    with open(full_path, "w") as outfile:
-        json.dump(json_obj, outfile, ensure_ascii=False, indent=2)
-
-def save_report_file(report, report_fields, csv_file, dir_name):
-    dir_path = os.path.join(DATA['DIR'],dir_name)
-    if not os.path.exists(dir_path):
-        os.mkdir(dir_path)
-    filepath =  os.path.join(dir_path, 'report.csv')
-    with open(filepath, 'w') as csvfile:
-            writer = csv.DictWriter(csvfile, fieldnames=report_fields)
-            writer.writeheader()
-            writer.writerows(report)
-    # save copy of input file
-    filepath =  os.path.join(dir_path, 'input.csv')
-    csv_file.seek(0)
-    with open(filepath, 'wb+') as f:
-        for chunk in csv_file.chunks():
-            f.write(chunk)
-
-def process_csv(csv_file, version):
-    print("Processing CSV")
-    dir_name = datetime.now().strftime("%Y-%m-%d-%H:%M:%S") + "-ror-records"
-    errors = None
-    success_msg = None
-    report = []
-    report_fields = ['row', 'ror_id', 'action', 'errors']
-    skipped_count = 0
-    updated_count = 0
-    new_count = 0
-    read_file = csv_file.read().decode('utf-8')
-    print(read_file)
-    reader = csv.DictReader(io.StringIO(read_file))
-    row_num = 2
-    for row in reader:
-        ror_id = None
-        updated = False
-        print("Row data")
-        print(row)
-        if row['id']:
-            ror_id = row['id']
-            updated = True
-            row_errors, v2_record = update_record_from_csv(row, version)
-        else:
-            row_errors, v2_record = new_record_from_csv(row, version)
-        if not row_errors:
-            if updated:
-                action = 'updated'
-                updated_count += 1
-            else:
-                action = 'created'
-                new_count += 1
-            ror_id = v2_record['id']
-            sorted_record = sort_list_fields(v2_record)
-            serializer = OrganizationSerializerV2(sorted_record)
-            json_obj = json.loads(JSONRenderer().render(serializer.data))
-            print(json_obj)
-            #create file
-            file = save_record_file(ror_id, updated, json_obj, dir_name)
-        else:
-            action = 'skipped'
-            skipped_count += 1
-            #print(errors)
-        report.append({"row": row_num, "ror_id": ror_id if ror_id else '', "action": action, "errors": row_errors if row_errors else ''})
-        row_num += 1
-        print(report)
-    if new_count > 0 or updated_count > 0 or skipped_count > 0:
-        #create report file
-        save_report_file(report, report_fields, csv_file, dir_name)
-        # create zip file
-        zipfile = shutil.make_archive(os.path.join(DATA['DIR'], dir_name), 'zip', DATA['DIR'], dir_name)
-        # upload to S3
-        '''
-        try:
-            DATA['CLIENT'].upload_file(zipfile, DATA['PUBLIC_STORE'], dir_name + '.zip')
-            zipfile = f"https://s3.eu-west-1.amazonaws.com/{DATA['PUBLIC_STORE']}/{urllib.parse.quote(dir_name)}.zip"
-        except Exception as e:
-            errors = e
-            print(errors)
-        '''
-
-    success_msg = {"file": zipfile, "rows processed": new_count + updated_count + skipped_count, "created": new_count, "udpated": updated_count, "skipped": skipped_count}
-    print(success_msg)
-    return success_msg
